api/requirements.txt
Flask
langchain
langchain_community
langchain_core
langchain-text-splitters
fastembed
pdfplumber
chromadb
PyYAML
waitress
Werkzeug
